pyproject.toml
[project]
name = "agents-assemble-discharge"
version = "0.1.0"
description = "Discharge Coordinator — A2A multi-agent discharge planning system"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.25.0",
    "a2a-sdk[http-server]>=0.3.26,<1.0.0",
    "litellm==1.83.7",
    "httpx>=0.27.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0,<3.0.0",
    "textstat>=0.7.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "pre-commit>=3.7.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "pre-commit>=3.7.0",
    "ruff>=0.15.12",
    "mypy>=2.0.0",
]
